Implement auth middleware

+++ backend/controllers/stuff.js
@@ -62,11 +62,31 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id }).then(
-        () => {
-            res.status(200).json({
-                message: 'Deleted!'
-            });
+    Thing.findOne({ _id: req.params.id }).then(
+        (thing) => {
+            if (!thing) {
+                return res.status(404).json({
+                    error: new Error('Thing not found!')
+                });
+            }
+            if (thing.userId !== req.auth.userId) {
+                return res.status(401).json({
+                    error: new Error('Unauthorized request!')
+                });
+            }
+            Thing.deleteOne({ _id: req.params.id }).then(
+                () => {
+                    res.status(200).json({
+                        message: 'Deleted!'
+                    });
+                }
+            ).catch(
+                (error) => {
+                    res.status(400).json({
+                        error: error
+                    });
+                }
+            );
         }
     ).catch(
         (error) => {
